Change MANIFEST reading function in pre-command hook
Replace plugin_read_config with plugin_read_list when reading the
manifest in the pre-command hook. This ensures list-version of the
option is fully supported.

Before this change the pre-command hook would complete with no error and
no cache restoration for file level cache when a manifest list is used.